requirements.txt
python-telegram-bot==13.15
redis==5.0.0
environs==9.5.0